LogandRCA agent/requirements.txt
neo4j
transformers
torch
faiss-cpu
GitPython
sentence-transformers
